Rewrite okta authenticator to capture errors and be class based.
It should be possible to use different authenticator deploys, thus the
class is rewritten to not directly rely on os.environ; from_env exists for
that pathway, and also provides a way to use env prefixing for lookups from
environment.

In that world, you'd just wind up using a mod-config python shim that imports
and instantiates an authenticator- either passing org/domain/apitoken, or loading
those from the environment.

Via that refactoring, the code will now fail if you don't supply the env variables
it requires (for the default/original deploy pattern).

Finally, this also has defensive decorators deployed to capture exceptions- radius
doesn't log them on it's own, thus this intercepts, logs, and then lets the exception
fly.