Remove the logic of post_auth, but allow for hooking.
I can definitely see a world where folks map okta grouping logic
into vlans/vpn groups (whatever)- but the current implentation hardcodes
assumptions that this is being used in an EAP context, and it expects people
to modify code inline rather than treating this as a framework.

Instead, add a check and invocation of a hook for any class deriviation; that
would allow people to shim this so that they can do okta stuff to then set
control-group.

For the example this code had- doing certain behaviours based on the user-
I'd argue that doesn't belong in this codepath- it's not related to Okta.

Either way, this structure allows for end users to implement their own, while
removing the EAP requirement.

...As to why EAP shouldn't be required in post-auth- because if you're testing
the inner-tunnel directly via radtest (to verify your auth/setup or code, and
then to verify EAP- the way to debug this), it gets in the way and breaks things.

Fact is, okta authenticator is just an authenticator for credentials- it shouldn't
be aware of tunneling leveled, and this commit fixes that encapsulation violation.