Make email domain an optional configurable, and rename it.
Specifically if it's okta org allowing multiple different email domains
for username (which is totally valid), the previous logic would break
that support.

Being this is no longer a required field, renamed it to `OKTA_DEFAULT_EMAIL_DOMAIN`
and adjust the code to only enforce it if a (basic) regex match against
the username doesn't look like an email address.