Refactor how hooks are exposed.
The conversion to a class was to allow this code to be overridden and
hooked- to do custom derivations while reusing the basic okta logic.

For example, a post_auth hook that maps vpn groups based on okta group
membership.

Injection methods have been added that can be used for taking an
arbitrary instance of this class and injecting the relevant hook shim's
into the modules namespace that rlm_python expects- additionally, doing
so in a way that ensures each method has errors properly captured/logged.

Finally, the stub post_auth method was removed- folks shouldn't be calling
an empty hook (which that was), they should instead invoke it only if there
is an implementation to invoke.